content/media/test/test_paused_after_ended.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: paused</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var videos = [];

function ended(evt) {
  var v = evt.target;
  is(v.paused, false, v._name + " must not be paused after end");
  if (AllFinished(videos)) {
    SimpleTest.finish();
  }
}

for (var i=0; i<gPausedAfterEndedTests.length; ++i) {
  var v = document.createElement('video');
  var test = gPausedAfterEndedTests[i];
  if (!v.canPlayType(test.type))
    continue;
  v.src = test.name;
  v._name = test.name;
  v._finished = false;
  v.load();
  is(v.paused, true,  v._name + " must be paused at start");

  v.play();
  is(v.paused, false, v._name + " must not be paused after play");

  v.addEventListener("ended", ended, false);
}

if (videos.length == 0) {
  todo(false, "No types supported");
} else {
  SimpleTest.waitForExplicitFinish();
}

</script>
</pre>
</body>
</html>
